requirements.txt
streamlit>=1.35.0
pyyaml>=6.0.1
pandas>=2.1.0
pyarrow>=14.0.0
google-cloud-bigquery>=3.17.0
google-genai>=1.0.0
python-dotenv>=1.0.1
matplotlib>=3.7.0
db-types>=0.3.0
